Improve data tests and abstract common code.
Use mockdate to mock the tests date so that it's predictable.
