+++ lib/helpers.js
@@ -1,8 +1,9 @@
 'use strict';
 
-const fs   = require('fs');
+const fs = require('fs');
 const path = require('path');
-const sri  = require('sri-toolbox');
+const semver = require('semver');
+const sri = require('sri-toolbox');
 const yaml = require('js-yaml');
 
 const CONFIG_FILE_PATH = path.resolve(__dirname, '../config/_config.yml');
@@ -45,9 +46,36 @@ function getConfigPath() {
     return CONFIG_FILE_PATH;
 }
 
+function generateDataJson() {
+    const config = getConfig();
+    const data = {
+        timestamp: new Date().toISOString(),
+        bootstrap: {},
+        fontawesome: {}
+    };
+
+    config.bootstrap.forEach((bootstrap) => {
+        const bootstrapVersion = bootstrap.version;
+
+        if (semver.satisfies(semver.coerce(bootstrapVersion), '<4')) {
+            data.bootstrap[bootstrapVersion] = {
+                css: bootstrap.stylesheet,
+                js: bootstrap.javascript
+            };
+        }
+    });
+
+    config.fontawesome.forEach((fontawesome) => {
+        data.fontawesome[fontawesome.version] = fontawesome.stylesheet;
+    });
+
+    return data;
+}
+
 module.exports = {
     getConfig,
     getConfigPath,
+    generateDataJson,
     theme: {
         selected: selectedTheme,
         fetch: getTheme
