+++ scripts/wp-plugin.js
@@ -2,34 +2,13 @@
 
 'use strict';
 
-const fs     = require('fs');
-const path   = require('path');
-const semver = require('semver');
+const fs = require('fs');
+const path = require('path');
 const helpers = require('../lib/helpers.js');
 
-const config  = helpers.getConfig();
 const OUT_DIR = path.resolve(__dirname, '../data');
 
-const data = {
-    timestamp: new Date(),
-    bootstrap: {},
-    fontawesome: {}
-};
-
-config.bootstrap.forEach((bootstrap) => {
-    const bootstrapVersion = bootstrap.version;
-
-    if (semver.satisfies(semver.coerce(bootstrapVersion), '<4')) {
-        data.bootstrap[bootstrapVersion] = {
-            css: bootstrap.stylesheet,
-            js: bootstrap.javascript
-        };
-    }
-});
-
-config.fontawesome.forEach((fontawesome) => {
-    data.fontawesome[fontawesome.version] = fontawesome.stylesheet;
-});
+const data = helpers.generateDataJson();
 
 if (!fs.existsSync(OUT_DIR)) {
     fs.mkdirSync(OUT_DIR);
