+++ test/test_helpers.js
@@ -7,11 +7,14 @@ process.env.ENABLE_CRAWLING = true;
 
 const assert     = require('assert');
 const htmlEncode = require('htmlencode').htmlEncode;
+const mockDate   = require('mockdate');
 const request    = require('request');
 const validator  = require('html-validator');
 const helpers    = require('../lib/helpers.js');
 
 let response = {};
+
+mockDate.set('03/05/2018');
 
 const CONTENT_TYPE_MAP = {
     css: 'text/css; charset=utf-8',
